move helpers to util, test for no file, add frame count to reader
