+++ mio/cli/process.py
@@ -2,137 +2,26 @@
 Command line interface for offline video pre-processing.
 """
 
-import contextlib
-import re
 import shutil
 from pathlib import Path
 from typing import List, Optional
 
 import click
-import cv2
-import pandas as pd
 
-from mio.io import VideoWriter
+from mio.io import VideoReader, VideoWriter
 from mio.logging import init_logger
 from mio.models.process import DenoiseConfig
 from mio.process.stitch import RecordingData, RecordingDataBundle
 from mio.process.video import crop_run, denoise_run
-from mio.utils import validate_frame_count_alignment, validate_video_metadata_match
+from mio.utils import (
+    DEFAULT_PROCESS_DIR,
+    extract_mismatch_details,
+    resolve_output_path,
+    validate_frame_count_alignment,
+    validate_video_metadata_match,
+)
 
 logger = init_logger("mio.cli.process")
-
-DEFAULT_PROCESS_DIR = "mio_process"
-
-
-def _format_missing_frame_ranges(missing_indices: List[int]) -> List[str]:
-    """Convert a sorted list of missing frame indices into readable ranges."""
-    if not missing_indices:
-        return []
-
-    ranges = []
-    start = missing_indices[0]
-    end = missing_indices[0]
-
-    for idx in missing_indices[1:]:
-        if idx == end + 1:
-            end = idx
-        else:
-            ranges.append(f"{start}-{end}" if start != end else str(start))
-            start = idx
-            end = idx
-
-    ranges.append(f"{start}-{end}" if start != end else str(start))
-    return ranges
-
-
-def _extract_mismatch_details(
-    video_path: Path, is_valid: bool, error_msg: Optional[str], csv_df: Optional[pd.DataFrame]
-) -> Optional[dict]:
-    """Extract detailed mismatch information from validation result."""
-    if is_valid:
-        return None
-
-    video_path_obj = Path(video_path)
-
-    # Parse error message to determine error type
-    if error_msg is None:
-        return {"error_type": "unknown"}
-
-    error_lower = error_msg.lower()
-    if "not found" in error_lower:
-        return {"error_type": "csv_not_found", "error_msg": error_msg}
-    elif "failed to read csv" in error_lower:
-        return {"error_type": "csv_read_error", "error_msg": error_msg}
-    elif "does not have 'reconstructed_frame_index'" in error_lower:
-        return {"error_type": "missing_column", "error_msg": error_msg}
-    elif "could not open video" in error_lower or "failed to read video" in error_lower:
-        return {"error_type": "video_error", "error_msg": error_msg}
-    elif "frame indices mismatch" in error_lower or "frames" in error_lower:
-        # Extract frame mismatch details from CSV DataFrame if available
-        if csv_df is not None and "reconstructed_frame_index" in csv_df.columns:
-            # Try to extract frame count from error message to avoid re-opening video file
-            # Error message format: "Frame indices mismatch: ... Video has {frame_count} frames"
-            video_frame_count = None
-            match = re.search(r"Video has (\d+) frames", error_msg)
-            if match:
-                with contextlib.suppress(ValueError, AttributeError):
-                    video_frame_count = int(match.group(1))
-
-            # If we couldn't extract from error message, open video file as fallback
-            if video_frame_count is None:
-                try:
-                    cap = cv2.VideoCapture(str(video_path_obj))
-                    if cap.isOpened():
-                        video_frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-                        cap.release()
-                except Exception:
-                    pass
-
-            if video_frame_count is not None:
-                unique_frame_indices = set(csv_df["reconstructed_frame_index"].unique())
-                csv_frame_count = len(unique_frame_indices)
-                expected_frame_indices = set(range(video_frame_count))
-                missing_indices = sorted(expected_frame_indices - unique_frame_indices)
-
-                return {
-                    "error_type": "frame_mismatch",
-                    "video_frame_count": video_frame_count,
-                    "csv_frame_count": csv_frame_count,
-                    "missing_count": len(missing_indices),
-                    "missing_indices": missing_indices[:20],
-                    "missing_ranges": _format_missing_frame_ranges(missing_indices),
-                }
-
-        # Fallback: return error message
-        return {"error_type": "frame_mismatch", "error_msg": error_msg}
-
-    return {"error_type": "unknown", "error_msg": error_msg}
-
-
-def resolve_output_path(
-    input_path: Path,
-    suffix: str,
-    output: str = DEFAULT_PROCESS_DIR,
-) -> Path:
-    """
-    Resolve the output path for process commands.
-    Treats output as a directory and generates filename with suffix.
-
-    Parameters:
-    input_path: Path to the input file
-    suffix: Suffix to add to filename (e.g., "_cropped", "_stitched")
-    output: Output directory (default: DEFAULT_PROCESS_DIR)
-
-    Returns:
-    Resolved output Path
-
-    .. todo::
-        Might be better to make this a genereic helper.
-        Putting it in cli for nowas we want this for process commands.
-    """
-    output_dir = Path(output).expanduser()
-    output_dir.mkdir(parents=True, exist_ok=True)
-    return output_dir / f"{input_path.stem}{suffix}{input_path.suffix}"
 
 
 @click.group()
@@ -493,9 +382,7 @@ def workflow(
         is_valid, error_msg, csv_df = validate_video_metadata_match(str(video_path_obj))
 
         if not is_valid:
-            mismatch_details = _extract_mismatch_details(
-                video_path_obj, is_valid, error_msg, csv_df
-            )
+            mismatch_details = extract_mismatch_details(video_path_obj, is_valid, error_msg, csv_df)
             validation_failures.append((video_path_obj, mismatch_details))
 
     # If there are validation failures, show dialogue
@@ -647,9 +534,9 @@ def workflow(
         click.echo("Trimming video...")
 
         # Get video frame count to calculate trim range
-        cap = cv2.VideoCapture(str(stitched_video))
-        total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-        cap.release()
+        reader = VideoReader(str(stitched_video))
+        total_frames = reader.frame_count
+        reader.release()
 
         crop_start = trim_start
         crop_end = total_frames - trim_end - 1  # -1 because trim_end is inclusive
