+++ mio/utils.py
@@ -2,15 +2,16 @@
 The junk drawer my dogs
 """
 
+import contextlib
 import hashlib
+import re
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional, Tuple, Union
+from typing import List, Optional, Tuple, Union
 
 import cv2
 import pandas as pd
 
-if TYPE_CHECKING:
-    pass
+DEFAULT_PROCESS_DIR = "mio_process"
 
 
 def hash_file(path: Union[Path, str]) -> str:
@@ -233,3 +234,96 @@ def validate_frame_count_alignment(
         )
 
     return True, None
+
+
+def format_missing_frame_ranges(missing_indices: List[int]) -> List[str]:
+    """Convert a sorted list of missing frame indices into readable ranges."""
+    if not missing_indices:
+        return []
+
+    ranges = []
+    start = missing_indices[0]
+    end = missing_indices[0]
+
+    for idx in missing_indices[1:]:
+        if idx == end + 1:
+            end = idx
+        else:
+            ranges.append(f"{start}-{end}" if start != end else str(start))
+            start = idx
+            end = idx
+
+    ranges.append(f"{start}-{end}" if start != end else str(start))
+    return ranges
+
+
+def extract_mismatch_details(
+    video_path: Path, is_valid: bool, error_msg: Optional[str], csv_df: Optional[pd.DataFrame]
+) -> Optional[dict]:
+    """Extract detailed mismatch information from validation result."""
+    if is_valid:
+        return None
+
+    video_path_obj = Path(video_path)
+
+    if error_msg is None:
+        return {"error_type": "unknown"}
+
+    error_lower = error_msg.lower()
+    if "not found" in error_lower:
+        return {"error_type": "csv_not_found", "error_msg": error_msg}
+    elif "failed to read csv" in error_lower:
+        return {"error_type": "csv_read_error", "error_msg": error_msg}
+    elif "does not have 'reconstructed_frame_index'" in error_lower:
+        return {"error_type": "missing_column", "error_msg": error_msg}
+    elif "could not open video" in error_lower or "failed to read video" in error_lower:
+        return {"error_type": "video_error", "error_msg": error_msg}
+    elif "frame indices mismatch" in error_lower or "frames" in error_lower:
+        if csv_df is not None and "reconstructed_frame_index" in csv_df.columns:
+            video_frame_count = None
+            match = re.search(r"Video has (\d+) frames", error_msg)
+            if match:
+                with contextlib.suppress(ValueError, AttributeError):
+                    video_frame_count = int(match.group(1))
+
+            if video_frame_count is None:
+                try:
+                    cap = cv2.VideoCapture(str(video_path_obj))
+                    if cap.isOpened():
+                        video_frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+                        cap.release()
+                except Exception:
+                    pass
+
+            if video_frame_count is not None:
+                unique_frame_indices = set(csv_df["reconstructed_frame_index"].unique())
+                csv_frame_count = len(unique_frame_indices)
+                expected_frame_indices = set(range(video_frame_count))
+                missing_indices = sorted(expected_frame_indices - unique_frame_indices)
+
+                return {
+                    "error_type": "frame_mismatch",
+                    "video_frame_count": video_frame_count,
+                    "csv_frame_count": csv_frame_count,
+                    "missing_count": len(missing_indices),
+                    "missing_indices": missing_indices[:20],
+                    "missing_ranges": format_missing_frame_ranges(missing_indices),
+                }
+
+        return {"error_type": "frame_mismatch", "error_msg": error_msg}
+
+    return {"error_type": "unknown", "error_msg": error_msg}
+
+
+def resolve_output_path(
+    input_path: Path,
+    suffix: str,
+    output: str = DEFAULT_PROCESS_DIR,
+) -> Path:
+    """
+    Resolve the output path for process commands.
+    Treats output as a directory and generates filename with suffix.
+    """
+    output_dir = Path(output).expanduser()
+    output_dir.mkdir(parents=True, exist_ok=True)
+    return output_dir / f"{input_path.stem}{suffix}{input_path.suffix}"
